requirements.txt
streamlit
pandas
faiss-cpu
openai
sentence-transformers
numpy
python-dotenv
scikit-learn
google-generativeai
